Re-factored InitializeCorrespondenceHandler to be smaller
